Add special-case hacks for BinData class
In the schema BinData is an extension of string, but modeling this in python
as a subclass of str with extra constructor args is not feasible. Instead
this change moves the string content to a `value` attribute which aligns
with how our to_dict converter presents element text content.

Fixes #9